Simplified internal data processing (class structure).
