--- src/main/java/de/dennisguse/opentracks/content/DataSource.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2011 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package de.dennisguse.opentracks.content;
-
-import android.content.ContentResolver;
-import android.content.Context;
-import android.database.ContentObserver;
-import android.net.Uri;
-
-/**
- * Data source on the phone.
- *
- * @author Rodrigo Damazio
- */
-class DataSource {
-
-    private final ContentResolver contentResolver;
-
-    DataSource(Context context) {
-        contentResolver = context.getContentResolver();
-    }
-
-    void registerContentObserver(Uri uri, ContentObserver observer) {
-        contentResolver.registerContentObserver(uri, false, observer);
-    }
-
-    void unregisterContentObserver(ContentObserver observer) {
-        contentResolver.unregisterContentObserver(observer);
-    }
-}
